chore(deps): update all dependencies to their latest versions
Bump dev tooling and runtime dependencies across the workspace:
- typescript 5.9 -> 6.0
- eslint 9 -> 10, @eslint/js 9 -> 10
- @types/node 25 -> 26
- tsdown 0.20 -> 0.22, typedoc 0.28.5 -> 0.28.19, vitest 4.0 -> 4.1
- @typescript-eslint/* 8.53 -> 8.61, eslint-config-turbo 2.7 -> 2.9
- @noble/curves & @noble/hashes 2.0 -> 2.2
- prettier, turbo and @changesets/cli to latest

Adjustments required by the upgrades:
- tsconfig: add the DOM lib (TextEncoder / Web Crypto globals are no longer
  provided as Node globals by @types/node 26); widen core's rootDir so
  TypeScript 6 accepts the interop test's cross-package source imports (TS6059).
- tsdown: migrate the deprecated `external` option to `deps.neverBundle`, and
  disable the ineffective-dynamic-import check in core (intentional pattern
  under code splitting).
- source: cast getRandomValues buffers to Uint8Array<ArrayBuffer> for the
  stricter lib.dom signature, attach `cause` to a re-thrown error, and drop
  type assertions now flagged as unnecessary by TS 6's improved inference.

format, lint, typecheck, build and tests all pass with no warnings.

+++ packages/secp256k1-tr/src/keys/index.ts
@@ -210,10 +210,10 @@ export function publicKeyPackageIntoEvenY(
 ): PublicKeyPackage {
   const evenY = isEven ?? publicKeyPackageHasEvenY(pkg);
   if (!evenY) {
-    const verifyingKey = Secp256K1Group.negate(pkg.verifyingKey as Uint8Array);
+    const verifyingKey = Secp256K1Group.negate(pkg.verifyingKey);
     const verifyingShares = new Map<string, Uint8Array>();
     for (const [id, share] of pkg.verifyingShares) {
-      verifyingShares.set(id, Secp256K1Group.negate(share as Uint8Array));
+      verifyingShares.set(id, Secp256K1Group.negate(share));
     }
     return {
       ...pkg,
@@ -239,10 +239,10 @@ export function publicKeyPackageTweak(
   const tp = Secp256K1Group.basePointMul(t);
 
   // Apply tweak to verifying key and all shares
-  const verifyingKey = Secp256K1Group.add(evenPkg.verifyingKey as Uint8Array, tp);
+  const verifyingKey = Secp256K1Group.add(evenPkg.verifyingKey, tp);
   const verifyingShares = new Map<string, Uint8Array>();
   for (const [id, share] of evenPkg.verifyingShares) {
-    verifyingShares.set(id, Secp256K1Group.add(share as Uint8Array, tp));
+    verifyingShares.set(id, Secp256K1Group.add(share, tp));
   }
 
   return {
@@ -269,9 +269,9 @@ export function keyPackageHasEvenY(pkg: KeyPackage): boolean {
 export function keyPackageIntoEvenY(pkg: KeyPackage, isEven?: boolean): KeyPackage {
   const evenY = isEven ?? keyPackageHasEvenY(pkg);
   if (!evenY) {
-    const verifyingKey = Secp256K1Group.negate(pkg.verifyingKey as Uint8Array);
-    const signingShare = Secp256K1ScalarField.negate(pkg.signingShare as bigint);
-    const verifyingShare = Secp256K1Group.negate(pkg.verifyingShare as Uint8Array);
+    const verifyingKey = Secp256K1Group.negate(pkg.verifyingKey);
+    const signingShare = Secp256K1ScalarField.negate(pkg.signingShare);
+    const verifyingShare = Secp256K1Group.negate(pkg.verifyingShare);
     return {
       ...pkg,
       verifyingKey,
@@ -294,9 +294,9 @@ export function keyPackageTweak(pkg: KeyPackage, merkleRoot?: Uint8Array): KeyPa
   const tp = Secp256K1Group.basePointMul(t);
 
   // Apply tweak to all components
-  const verifyingKey = Secp256K1Group.add(evenPkg.verifyingKey as Uint8Array, tp);
-  const signingShare = Secp256K1ScalarField.add(evenPkg.signingShare as bigint, t);
-  const verifyingShare = Secp256K1Group.add(evenPkg.verifyingShare as Uint8Array, tp);
+  const verifyingKey = Secp256K1Group.add(evenPkg.verifyingKey, tp);
+  const signingShare = Secp256K1ScalarField.add(evenPkg.signingShare, t);
+  const verifyingShare = Secp256K1Group.add(evenPkg.verifyingShare, tp);
 
   return {
     ...evenPkg,
